refactor: api와 dto파일 하위 폴더로 분류
