+++ post_fastapi/api/post/post.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from typing import List
 
-from api.post_request import PostUpdate
+from api.post.post_request import PostUpdate
 from domain.post import Post
 from fastapi import APIRouter, Depends, Response, status
 from fastapi.encoders import jsonable_encoder
@@ -12,13 +12,9 @@ from tool.session import verify_session
 router = APIRouter(prefix="/posts")
 
 
-@router.get("/healthcheck")
-async def root(post_service: PostService = Depends()):
-    print(post_service)
-    return Response(content="ok", status_code=status.HTTP_200_OK)
-
-
-@router.get("", status_code=status.HTTP_200_OK, response_model=List[Post])
+@router.get(
+    "", status_code=status.HTTP_200_OK, response_model=List[Post]
+)  # 파라미터의 순서 일관된게 맞추기
 async def get_posts(post_service: PostService = Depends()):
     """
     게시글id와 게시글 전체를 반환합니다.
@@ -34,18 +30,18 @@ async def get_posts(post_service: PostService = Depends()):
 @router.get("/{id}", response_model=Post, status_code=status.HTTP_200_OK)
 async def get_post(id: int, post_service: PostService = Depends()):
     """
-    id 값에 해당하는 게시물 반환합니다.
+        id 값에 해당하는 게시물 반환합니다.
 
-    Parameters :
-        id : 게시물의 id
+        Parameters :
+            id : 게시물의 id
 
-    Return :
-        Post :
-            user: 작성자
-            title: 제목
-            content : 글 내용
-            create_date : 생성 시간
-
+        Return :
+            Post :
+                user: 작성자
+                title: 제목
+                content : 글 내용
+                create_date : 생성 시간
+    #!의미 없는 개행 삭제
 
     """
     post = post_service.get_post(id)
@@ -57,7 +53,7 @@ async def get_post(id: int, post_service: PostService = Depends()):
 @router.post("", response_model=Post, status_code=status.HTTP_201_CREATED)
 async def create(
     post: Post,
-    user: str = Depends(verify_session),
+    user: str = Depends(verify_session),  #! user, user_id 이름 맞추기
     post_service: PostService = Depends(),
 ):
     """
@@ -79,7 +75,6 @@ async def create(
 
     """
     post.user = user.username
-    print(post)
     post = post_service.create_post(post)
 
     return JSONResponse(
@@ -107,7 +102,7 @@ async def update_post(id: int, post: PostUpdate, post_service: PostService = Dep
 
 
     """
-    post = post_service.update_post(id, post)
+    post = post_service.update_post(id, post)  # 해당 유저만 수정가능하게 수정()
 
     return JSONResponse(content=jsonable_encoder(post), status_code=status.HTTP_200_OK)
