--- post_fastapi/api/post_dto.py
@@ -1,17 +1,0 @@
-from datetime import datetime
-from typing import Optional
-
-from sqlmodel import Field, SQLModel
-
-
-class PostUpdate(SQLModel):
-    user: Optional[str]
-    title: Optional[str]
-    content: Optional[str]
-    updated_at: datetime = Field(default_factory=datetime.today)
-
-
-class PostRequest(SQLModel):
-    user: Optional[str]
-    title: Optional[str]
-    content: Optional[str]
